python/requirements.txt
SQLAlchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
